requirements.txt
decorator==4.1.2
networkx==2.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
